requirements/inference.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/inference.txt requirements/inference.in
#
aistudio-sdk==0.3.8
    # via paddlex
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via httpx
astor==0.8.1
    # via paddlepaddle-gpu
bce-python-sdk==0.9.56
    # via aistudio-sdk
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via paddlex
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   aistudio-sdk
    #   typer-slim
colorama==0.4.6
    # via
    #   click
    #   colorlog
    #   tqdm
colorlog==6.10.1
    # via paddlex
decorator==5.2.1
    # via paddlepaddle-gpu
exceptiongroup==1.3.1
    # via anyio
filelock==3.20.1
    # via
    #   huggingface-hub
    #   modelscope
    #   paddlex
fsspec==2025.12.0
    # via huggingface-hub
future==1.0.0
    # via bce-python-sdk
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   paddlepaddle-gpu
huggingface-hub==1.2.3
    # via paddlex
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via paddlex
modelscope==1.33.0
    # via paddlex
numpy==2.2.6
    # via
    #   opencv-contrib-python
    #   opencv-python
    #   opt-einsum
    #   paddlepaddle-gpu
    #   paddlex
    #   pandas
    #   shapely
opencv-contrib-python==4.10.0.84
    # via paddlex
opencv-python==4.12.0.88
    # via -r requirements/inference.in
opt-einsum==3.3.0
    # via paddlepaddle-gpu
packaging==25.0
    # via
    #   huggingface-hub
    #   paddlex
paddleocr==3.3.2
    # via -r requirements/inference.in
paddlepaddle-gpu==2.6.2
    # via -r requirements/inference.in
paddlex[ocr-core]==3.3.12
    # via paddleocr
pandas==2.3.3
    # via paddlex
pillow==12.0.0
    # via
    #   paddlepaddle-gpu
    #   paddlex
prettytable==3.17.0
    # via
    #   aistudio-sdk
    #   paddlex
protobuf==3.20.2
    # via paddlepaddle-gpu
psutil==7.1.3
    # via aistudio-sdk
py-cpuinfo==9.0.0
    # via paddlex
pyclipper==1.4.0
    # via paddlex
pycryptodome==3.23.0
    # via bce-python-sdk
pydantic==2.12.5
    # via paddlex
pydantic-core==2.41.5
    # via pydantic
pypdfium2==5.2.0
    # via paddlex
python-bidi==0.6.7
    # via paddlex
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   paddleocr
    #   paddlex
requests==2.32.5
    # via
    #   aistudio-sdk
    #   modelscope
    #   paddleocr
    #   paddlex
ruamel-yaml==0.18.17
    # via paddlex
ruamel-yaml-clib==0.2.15
    # via ruamel-yaml
shapely==2.1.2
    # via paddlex
shellingham==1.5.4
    # via huggingface-hub
six==1.17.0
    # via
    #   bce-python-sdk
    #   python-dateutil
tqdm==4.67.1
    # via
    #   aistudio-sdk
    #   huggingface-hub
    #   modelscope
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   huggingface-hub
    #   paddleocr
    #   paddlex
    #   pydantic
    #   pydantic-core
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
ujson==5.11.0
    # via paddlex
urllib3==2.6.2
    # via
    #   modelscope
    #   requests
wcwidth==0.2.14
    # via prettytable

# The following packages are considered to be unsafe in a requirements file:
# setuptools
